+++ R/model.R
@@ -1,11 +1,13 @@
 model <- function(vals, vars, method = "bioclim") {
 
-  # version 1.0 (31 Dec 2024)
+  # version 1.1 (2 Jan 2025)
 
   method <- tolower(method)
   method <- match.arg(method, choices = c("bioclim", "domain", "convexhull", "mahalanobis", "kernel", "mvnormal"))
 
   vals <- as.data.frame(vals)  # else errors for some methods if only one variable
+  if (inherits(vars, "SpatRaster")) vars <- vars[[names(vals)]]
+  else vars <- vars[ , names(vals), drop = FALSE]
 
   if (method == "bioclim")
     bioclim(vals, vars)
@@ -32,8 +34,17 @@ model <- function(vals, vars, method = "bioclim") {
   else if (method == "mahalanobis")
     mahalanobisModel(vals, vars)
 
-  else if (method == "kernel")
-    kernelModel(vals, vars)
+  else if (method == "kernel") {
+    # if (ncol(vals) > 6)
+    #   stop ("Kernel density estimate is only implemented for up to 6 variables.")
+    if (ncol(vals) > 4) {
+      message ("Setting binned=FALSE, as required by ks::kde() when >4 variables.\nThis is computationally intensive!")
+      # otherwise:
+      # Error in ks::kde(as.matrix(vals), compute.cont = FALSE, approx.cont = TRUE) : Binned estimation for d>4 not implemented. Set binned=FALSE for exact estimation.
+      kernelModel(vals, vars, binned = FALSE)
+    }
+    else kernelModel(vals, vars)
+  }
 
   else if (method == "mvnormal") {
     if (ncol(vals) < 2) stop("input 'method' requires more than one variable.")