+++ NEWS.md
@@ -1,3 +1,10 @@
+# proniche 0.2
+
+-   implemented for data frame (not just SpatRaster) variables
+
+-   added some more checks to prevent errors, e.g. due to wrong number of variables for method
+
+
 # proniche 0.1
 
 -   added package files (DESCRIPTION, NAMESPACE, NEWS; 'man' and 'R' folders)